fix: rename app
