+++ src/main.js
@@ -1,5 +1,5 @@
 import { createApp } from 'vue'
-import App from './App.vue'
+import App from './Hive.vue'
 import router from './router'
 
-createApp(App).use(router).mount('#app')
+createApp(App).use(router).mount('#Hive')
