+++ src/router/index.js
@@ -1,16 +1,16 @@
 import { createRouter, createWebHistory } from 'vue-router'
-import Home from '@/views/AppHome.vue'
-import About from '@/views/AppAbout.vue'
+import Home from '@/views/HiveHome.vue'
+import About from '@/views/HiveAbout.vue'
 
 const routes = [
   {
     path: '/',
-    name: 'AppHome',
+    name: 'HiveHome',
     component: Home
   },
   {
     path: '/about',
-    name: 'AppAbout',
+    name: 'HiveAbout',
     component: About
   }
 ]
