+++ src/views/HiveAbout.vue
@@ -9,7 +9,7 @@
 
 <script>
 export default {
-  name: 'AppAbout'
+  name: 'HiveAbout'
 }
 </script>
